Statistical analysis of precipitation data with Python
Usually we use probabilistic approaches when dealing with extreme
events since the size of available data is scarce to address the
maximum for a determined return period.
Precipitation data present challenges when we try to fit to a
statistical distribution. This is an exercise done with 10 years record
of precipitation for the Morropon station (Piura-Peru) affected by El
Niño and La Niña events in years 98 and 99 respectively (download data
here).
Data processing was done with Python on the IPython Notebook interface.
See the whole description in this link:
http://hatarilabs.tumblr.com/post/136633361772/statistical-analysis-of-p
recipitation-data-with